requirements.txt
appnope==0.1.3
asttokens==2.2.1
attrs==23.2.0
backcall==0.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
cmd2==1.5.0
colorama==0.4.6
comm==0.1.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
google-api-core==2.17.1
google-auth==2.28.1
google-cloud==0.34.0
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.2
google-cloud-speech==2.25.0
google-cloud-storage==2.14.0
google-cloud-texttospeech==2.16.2
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
grpcio-status==1.62.0
grpcio-tools==1.62.0
idna==3.6
jedi==0.18.2
jsonpath-ng==1.6.1
kuksa-client==0.4.2
nest-asyncio==1.5.6
numpy==1.26.4
packaging==23.1
pandas==2.2.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.1
ply==3.11
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2024.1
pyzmq==25.1.0
requests==2.31.0
rsa==4.9
shapely==2.0.3
six==1.16.0
SpeechRecognition==3.10.1
stack-data==0.6.2
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.6
websockets==12.0
